# GGRC-7242 Improve performance for mapping big program