@mrR2D2 2019-07-01
> Have you measured performance on perf instance before and after changes? please put them to PR description

@belovmd updated description with timing measurements I got.